Moved source files to lib folder
