+++ lib/generateExampleJson.js
