+++ lib/getSchemaFromObject.js
