+++ lib/getSchemasDirectory.js
