+++ lib/parseExampleJson.js
