+++ lib/readRawSchemas.js
