+++ lib/schemaFactory.js
