+++ test/integration/dirCacheTest/test1/requireSchemagic.js
@@ -1,2 +1,2 @@
 delete require.cache[__filename]; //do not cache in require cache
-module.exports = require('../../../../');
+module.exports = require('../../../../lib');
