+++ test/integration/dirCacheTest/test2/test2subdir/requireSchemagic.js
@@ -1,2 +1,2 @@
 delete require.cache[__filename]; //do not cache in require cache
-module.exports = require('../../../../../');
+module.exports = require('../../../../../lib');
